oops

+++ semester6/CG/lab1/Shape.hpp
@@ -47,7 +47,7 @@ struct Shape
             const Point& p2 = edge.p2;
             if (p1 == point || p2 == point) return true;
             if (p1.y > point.y || p2.y < point.y) continue;
-            int num = (p1.y - p2.y) * point.x + (p2.x - p1.x) * point.y + (p1.x * p2.y - p2.x * p1.y);
+            double num = (p1.y - p2.y) * point.x + (p2.x - p1.x) * point.y + (p1.x * p2.y - p2.x * p1.y);
             if (num <= 0) ++res;
         }
         return (res % 2 != 0);
